refactor(cli): move cross-command helpers into cli/_shared.py
Relocate _looks_like_sponsio_config, _resolve_entry, _parse_since, and
_parse_existing_contracts (each used by 2+ commands, and _resolve_entry
imported by eval_runner) into _shared.py alongside _contract_guarantee.
__init__ re-exports _resolve_entry from there. No behavior change.
